rename tsx pages
