+++ src/pages/LandingPage.tsx
@@ -6,7 +6,7 @@ import LandingRegisterPrompt from '../components/LandingRegisterPrompt';
 import LandingInfo from '../components/LandingInfo';
 import Footer from '../components/Footer';
 
-const Landing: React.FC = () => {
+const LandingPage: React.FC = () => {
     return (
         <IonPage>
             <IonContent fullscreen>
@@ -22,4 +22,4 @@ const Landing: React.FC = () => {
     );
 };
 
-export default Landing;
+export default LandingPage;
